Fixed ban, David please pull

+++ src/commands/moderation/ban.js
@@ -6,7 +6,7 @@ module.exports = {
     category: 'Moderation',
     permissions: ['BAN_MEMBERS'],
     async execute(client, msg, args) {
-        const user = msg.mentions.members.first() || msg.guild.members.cache.get(args[0]) || msg.guild.members.cache.find(r => r.user.username === args[0]) || msg.guild.members.cache.find(ro => ro.displayName === args[0]);
+        const user = msg.mentions.members.first() || msg.guild.members.cache.get(args[0]) || args[0]
 
         if (user) {
             const bans = await msg.guild.fetchBans();
@@ -25,17 +25,16 @@ module.exports = {
             if (ban) {
                 return msg.channel.send('User is already banned');
             }
-            await user.send(`You've been banned from \`${msg.guild.name}\` for the reason: \`${args.slice(1).join(' ') || 'No reason provided.'}\`.`)
-                .catch(() => {
-                    msg.channel.send(`I was unable to DM ${user.user.tag}.`);
-                });
-            msg.channel.messages.fetch({
-                limit: 120
-            }).then((messages) => {
-                const Messages = [];
-                messages.filter(m => m.author.id === user.user.id).forEach(msg => Messages.push(msg))
-                msg.channel.bulkDelete(Messages)
-            })
+
+            if(args.join(" ").toLowerCase().includes(`-c`)) {
+                msg.channel.messages.fetch({
+                    limit: 100
+                }).then((messages) => {
+                    const Messages = [];
+                    messages.filter(m => m.author.id === user.user.id).forEach(msg => Messages.push(msg))
+                    msg.channel.bulkDelete(Messages)
+                })
+            }
 
             msg.guild.members.ban(user, {
                 reason: `${args.slice(1).join(' ') || 'No reason provided'} [${msg.author.tag}}`
